requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.14
django-ckeditor==6.7.1
django-environ==0.11.2
django-js-asset==2.2.0
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pillow==10.4.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sqlparse==0.5.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
